+++ src/libstddjb/ip46_scanlist.c
@@ -1,17 +1,17 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/ip46.h>
-
 #include <skalibs/bytestr.h>
 #include <skalibs/fmtscan.h>
 
-unsigned int ip46full_scanlist (ip46full_t *out, unsigned int max, char const *s, unsigned int *num)
+size_t ip46full_scanlist (ip46full_t *out, size_t max, char const *s, size_t *num)
 {
-  unsigned int n = 0, w = 0 ;
+  size_t n = 0, w = 0 ;
   for (; s[w] && (n < max) ; n++)
   {
     ip46full_t z ;
-    register unsigned int i = ip6_scan(s + w, z.ip) ;
+    register size_t i = ip6_scan(s + w, z.ip) ;
     if (i) z.is6 = 1 ;
     else
     {
