+++ src/libstddjb/ip6_scanlist.c
@@ -1,15 +1,16 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bytestr.h>
 #include <skalibs/fmtscan.h>
 
-unsigned int ip6_scanlist (char *out, unsigned int max, char const *s, unsigned int *num)
+size_t ip6_scanlist (char *out, size_t max, char const *s, size_t *num)
 {
-  unsigned int n = 0, w = 0 ;
+  size_t n = 0, w = 0 ;
   for (; s[w] && (n < max) ; n++)
   {
     char ip[16] ;
-    register unsigned int i = ip6_scan(s + w, ip) ;
+    register size_t i = ip6_scan(s + w, ip) ;
     if (!i) break ;
     byte_copy(out + (n << 4), 16, ip) ;
     w += i ;
