+++ src/libstddjb/localtmn_fmt.c
@@ -1,9 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/uint32.h>
 #include <skalibs/djbtime.h>
 
-unsigned int localtmn_fmt (char *s, localtmn_t const *l)
+size_t localtmn_fmt (char *s, localtmn_t const *l)
 {
   char *p = s ;
   p += localtm_fmt(p, &l->tm) ; *p++ = '.' ;
