+++ src/libstddjb/uint16_reverse.c
@@ -1,8 +1,9 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/uint16.h>
 
-void uint16_reverse (char *s, unsigned int n)
+void uint16_reverse (char *s, size_t n)
 {
   while (n--)
   {
