+++ src/libstddjb/uint32_reverse.c
@@ -1,8 +1,9 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/uint32.h>
 
-void uint32_reverse (char *s, unsigned int n)
+void uint32_reverse (char *s, size_t n)
 {
   while (n--)
   {
